requirements.txt
mysql-connector-python
requests
psycopg2
pandas
sqlalchemy
apache-airflow
python-dotenv